minimaze changes
